See description
Added the isGood function to also parse out the wiki ‘good’ article
stamp.  Implemented this in both the titleSearch and read_file
functions.  The main function is set up to allow for user input to
search for title.  Need to allow for multi-word inputs though. Also
added my xml data dump to the ignore folder.