Added picture count etc.
Added the picCount function to return the number of times .jpg is found
in the article. Added short pic member in wikiPage class to hold pic
count and included a call to picCount in the wikiPage constructor to
initialize the picture count.

Also added a removeJunk function meant to remove some of the useless
formatting of the xml files (not done)